+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## Version 1.32
-* issue 6563: export data from MSP
+* issue 6563: export data from MSP and fix address on export for external researcher
 
 ## Version 1.31
 * issue 7505: generate study participation ID in the invite task before sending out the actual email.